Update Exploratory Analysis with Life Stages.Rmd
Hi Sebastian,

It might be a good idea to put descriptive titles within code chunks to explain what each chunk is doing/ what graph it produces. It helps provide clarity on what the code is, so people don't have to read each line of code before understanding what it is doing entirely. As an example, see line 31 where I added that title.

Also, there are some stats stuff within the 'Modifying and plotting data' heading (ex line 94-98). Depending on flow of ideas, it might make more sense to keep it where it is or move it to the Stats heading that appears later. You could always make subheadings for statistics within the 'Modifying and plotting data' heading. Your choice.

Otherwise, all the code is clear and concise.